change password
Seed and migration made minor changes

+++ app/Traits/FileUploadTrait.php
@@ -6,14 +6,15 @@ use Illuminate\Http\Request;
 
 trait FileUploadTrait
 {
-    public function uploadImage(Request $request ,string $inputName ,string $oldPath = Null ,string $prePath = '/uploads'): ?string
+    public function uploadImage(Request $request ,string $inputName ,?string $oldPath = Null ,string $prePath = '/uploads'): ?string
     {
         if ($request->hasFile($inputName)){
             $image = $request->{$inputName};
             $ext = $image->getClientOriginalExtension();
             $imagName = 'media'.uniqid().'.'.$ext ;
             $image->move(public_path($prePath),$imagName);
-            if ($oldPath && File::exists(public_path($oldPath))){
+            $exceptPath = '/default';
+            if ($oldPath && File::exists(public_path($oldPath)) && strpos($oldPath, $exceptPath) !== 0 ){
                 File::delete(public_path($oldPath));
             }
             return $prePath.'/'.$imagName ;
